[Created] - Create RumContactUsView to encapsulate layout responsabilities
