+++ CleanCodeApp/Modules/Features/Rum/ContactsUs/RumContactUsViewController.swift
@@ -7,165 +7,31 @@
 
 import UIKit
 
-class RumContactUsViewController: LoadingInheritageController {
+final class RumContactUsViewController: LoadingInheritageController {
     // MARK: - Model
     var model: ContactUsModel?
     
     // MARK: - UI Components
-    private lazy var textView: UITextView = {
-        let textView = UITextView()
-        textView.text = "Escreva sua mensagem aqui"
-        textView.translatesAutoresizingMaskIntoConstraints = false
-        return textView
-    }()
-    
-    private lazy var titleLabel: UILabel = {
-        let titleLabel = UILabel()
-        titleLabel.textColor = .black
-        titleLabel.font = UIFont.systemFont(ofSize: 24, weight: .semibold)
-        titleLabel.text = "Escolha o canal para contato"
-        titleLabel.translatesAutoresizingMaskIntoConstraints = false
-        return titleLabel
-    }()
-    
-    private lazy var phoneButton: UIButton = {
-        let phoneButton = UIButton()
-        phoneButton.backgroundColor = .systemGray4
-        phoneButton.layer.cornerRadius = 10
-        phoneButton.addTarget(self, action: #selector(didTapPhoneButton), for: .touchUpInside)
-        phoneButton.translatesAutoresizingMaskIntoConstraints = false
-        return phoneButton
-    }()
-    
-    private lazy var emailButton: UIButton = {
-        let emailButton = UIButton()
-        emailButton.backgroundColor = .systemGray4
-        emailButton.layer.cornerRadius = 10
-        emailButton.addTarget(self, action: #selector(didTapEmailButton), for: .touchUpInside)
-        emailButton.translatesAutoresizingMaskIntoConstraints = false
-        return emailButton
-    }()
-    
-    private lazy var chatButton: UIButton = {
-        let chatButton = UIButton()
-        chatButton.backgroundColor = .systemGray4
-        chatButton.layer.cornerRadius = 10
-        chatButton.addTarget(self, action: #selector(didTapChatButton), for: .touchUpInside)
-        chatButton.translatesAutoresizingMaskIntoConstraints = false
-        return chatButton
-    }()
-    
-    private lazy var messageLabel: UILabel = {
-        let messageLabel = UILabel()
-        messageLabel.textColor = .black
-        messageLabel.font = UIFont.systemFont(ofSize: 16, weight: .semibold)
-        messageLabel.text = "Ou envie uma mensagem"
-        messageLabel.numberOfLines = 2
-        messageLabel.setContentHuggingPriority(.defaultLow, for: .horizontal)
-        messageLabel.translatesAutoresizingMaskIntoConstraints = false
-        
-        return messageLabel
-    }()
-    
-    private lazy var sendMessageButton: UIButton = {
-        let sendMessageButton = UIButton()
-        sendMessageButton.backgroundColor = .blue
-        sendMessageButton.setTitle("  Enviar ", for: .normal)
-        sendMessageButton.setTitleColor(.white, for: .normal)
-        sendMessageButton.layer.cornerRadius = 10
-        sendMessageButton.setContentHuggingPriority(.required, for: .horizontal)
-        sendMessageButton.addTarget(self, action: #selector(didTapSendMessageButton), for: .touchUpInside)
-        sendMessageButton.translatesAutoresizingMaskIntoConstraints = false
-        return sendMessageButton
-    }()
-    
-    private lazy var closeButton: UIButton = {
-        let closeButton = UIButton()
-        closeButton.setTitle("Voltar", for: .normal)
-        closeButton.setTitleColor(.blue, for: .normal)
-        closeButton.backgroundColor = .clear
-        closeButton.layer.borderWidth = 1
-        closeButton.layer.borderColor = UIColor.blue.cgColor
-        closeButton.layer.cornerRadius = 10
-        closeButton.addTarget(self, action: #selector(didTapCloseButton), for: .touchUpInside)
-        closeButton.translatesAutoresizingMaskIntoConstraints = false
-        return closeButton
-    }()
+    private lazy var contactUsView = RumContactUsView()
     
     // MARK: - Lifecycle
+    override func loadView() {
+        view = contactUsView
+    }
+    
     override func viewDidLoad() {
         super.viewDidLoad()
-        setupView()
+        setupActions()
         fetchData()
     }
     
     // MARK: - Setup Methods
-    private func setupView() {
-        view.backgroundColor = .systemGray6
-        setupButtonImages()
-        setupHierarchy()
-        setupConstraints()
-    }
-    
-    private func setupButtonImages() {
-        phoneButton.setSymbolImage(systemName: "phone", pointSize: 36)
-        emailButton.setSymbolImage(systemName: "envelope", pointSize: 36)
-        chatButton.setSymbolImage(systemName: "message", pointSize: 36)
-    }
-    
-    private func setupHierarchy() {
-        view.addSubview(titleLabel)
-        view.addSubview(phoneButton)
-        view.addSubview(emailButton)
-        view.addSubview(chatButton)
-        view.addSubview(messageLabel)
-        view.addSubview(textView)
-        view.addSubview(sendMessageButton)
-        view.addSubview(closeButton)
-    }
-    
-    private func setupConstraints() {
-        NSLayoutConstraint.activate([
-            titleLabel.topAnchor.constraint(equalTo: view.safeAreaLayoutGuide.topAnchor, constant: 30),
-            titleLabel.leadingAnchor.constraint(equalTo: view.leadingAnchor, constant: 20),
-            titleLabel.trailingAnchor.constraint(equalTo: view.trailingAnchor, constant: -20),
-            
-            phoneButton.topAnchor.constraint(equalTo: titleLabel.bottomAnchor, constant: 30),
-            emailButton.centerYAnchor.constraint(equalTo: phoneButton.centerYAnchor),
-            chatButton.centerYAnchor.constraint(equalTo: phoneButton.centerYAnchor),
-            
-            phoneButton.widthAnchor.constraint(equalToConstant: 80),
-            phoneButton.heightAnchor.constraint(equalToConstant: 80),
-            phoneButton.leadingAnchor.constraint(equalTo: view.leadingAnchor, constant: 20),
-            
-            emailButton.widthAnchor.constraint(equalToConstant: 80),
-            emailButton.heightAnchor.constraint(equalToConstant: 80),
-            emailButton.centerXAnchor.constraint(equalTo: view.centerXAnchor),
-            
-            chatButton.widthAnchor.constraint(equalToConstant: 80),
-            chatButton.heightAnchor.constraint(equalToConstant: 80),
-            chatButton.trailingAnchor.constraint(equalTo: view.trailingAnchor, constant: -20),
-            
-            messageLabel.topAnchor.constraint(equalTo: phoneButton.bottomAnchor, constant: 30),
-            messageLabel.leadingAnchor.constraint(equalTo: view.leadingAnchor, constant: 20),
-            messageLabel.trailingAnchor.constraint(equalTo: view.trailingAnchor, constant: -20),
-            //            stackView.heightAnchor.constraint(equalToConstant: 30),
-            
-            textView.topAnchor.constraint(equalTo: messageLabel.bottomAnchor, constant: 20),
-            textView.leadingAnchor.constraint(equalTo: view.leadingAnchor, constant: 20),
-            textView.trailingAnchor.constraint(equalTo: view.trailingAnchor, constant: -20),
-            textView.bottomAnchor.constraint(equalTo: sendMessageButton.topAnchor, constant: -30),
-            
-            sendMessageButton.bottomAnchor.constraint(equalTo: closeButton.topAnchor, constant: -20),
-            sendMessageButton.heightAnchor.constraint(equalToConstant: 40),
-            sendMessageButton.leadingAnchor.constraint(equalTo: view.leadingAnchor, constant: 20),
-            sendMessageButton.trailingAnchor.constraint(equalTo: view.trailingAnchor, constant: -20),
-            
-            closeButton.bottomAnchor.constraint(equalTo: view.safeAreaLayoutGuide.bottomAnchor, constant: -20),
-            closeButton.heightAnchor.constraint(equalToConstant: 40),
-            closeButton.leadingAnchor.constraint(equalTo: view.leadingAnchor, constant: 20),
-            closeButton.trailingAnchor.constraint(equalTo: view.trailingAnchor, constant: -20),
-        ])
+    private func setupActions() {
+        contactUsView.phoneButton.addTarget(self, action: #selector(didTapPhoneButton), for: .touchUpInside)
+        contactUsView.emailButton.addTarget(self, action: #selector(didTapEmailButton), for: .touchUpInside)
+        contactUsView.chatButton.addTarget(self, action: #selector(didTapChatButton), for: .touchUpInside)
+        contactUsView.sendMessageButton.addTarget(self, action: #selector(didTapSendMessageButton), for: .touchUpInside)
+        contactUsView.closeButton.addTarget(self, action: #selector(didTapCloseButton), for: .touchUpInside)
     }
     
     // MARK: - Actions
@@ -225,7 +91,7 @@ class RumContactUsViewController: LoadingInheritageController {
     @objc func didTapSendMessageButton() {
         view.endEditing(true)
         let email = model?.mail ?? ""
-        if let message = textView.text, textView.text.count > 0 {
+        if let message = contactUsView.textView.text, contactUsView.textView.text.count > 0 {
             let parameters: [String: String] = [
                 "email": email,
                 "mensagem": message
